# server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ECOMMERCE MERN SHOP",
    "description": "API for the MERN stack ECOMMERCE app by <b>David Enoragbon</b> using NodeJS."
  },
  "host": "localhost:5000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "Authorization",
      "description": "Value: Bearer ",
      "in": "header",
      "scheme": "bearer"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/refresh_token": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/upload": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UploadModel"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/api/destroy": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeleteUploadModel"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "UploadModel": {
      "type": "object",
      "properties": {
        "file": {
          "type": "string",
          "example": "myImage.png"
        }
      },
      "required": [
        "file"
      ]
    },
    "DeleteUploadModel": {
      "type": "object",
      "properties": {
        "public_id": {
          "type": "string",
          "example": "ecommerce/xwir6kmuqvvrxvvyk9oh"
        }
      },
      "required": [
        "public_id"
      ]
    }
  }
}